// packages/onchainkit/package/src/transaction/types.ts
export type Call = { to: Hex; data?: Hex; value?: bigint };

type TransactionButtonOverride = {
  text?: ReactNode;
  onClick?: (receipt?: TransactionReceipt) => void;
};

/**
 * List of transaction lifecycle statuses.
 * The order of the statuses loosely follows the transaction lifecycle.
 *
 * Note: exported as public Type
 */
export type LifecycleStatus =
  | {
      statusName: 'init';
      statusData: null;
    }
  | {
      statusName: 'error';
      statusData: TransactionError;
    }
  | {
      statusName: 'transactionIdle'; // initial status prior to the mutation function executing
      statusData: null;
    }
  | {
      statusName: 'buildingTransaction';
      statusData: null;
    }
  | {
      statusName: 'transactionPending'; // if the mutation is currently executing
      statusData: null;
    }
  | {
      statusName: 'transactionLegacyExecuted';
      statusData: {
        transactionHashList: Address[];
      };
    }
  | {
      statusName: 'success'; // if the last mutation attempt was successful
      statusData: {
        transactionReceipts: TransactionReceipt[];
      };
    }
  | {
      statusName: 'reset';
      statusData: null;
    };
// ...
type PaymasterService = {
  url: string;
};
// ...
export type TransactionDefaultReact = {
  disabled?: boolean;
} & Omit<TransactionReact, 'children'>;

export type Calls = Call[] | Promise<Call[]> | (() => Promise<Call[]>);
export type Contracts =
  | ContractFunctionParameters[]
  | Promise<ContractFunctionParameters[]>
  | (() => Promise<ContractFunctionParameters[]>);

export type TransactionProviderReact = {
  /** An array of calls to be made in the transaction */
  calls?: Calls | Contracts | Array<Call | ContractFunctionParameters>;
  /**
   * @deprecated Use `isSponsored` instead.
   */
  capabilities?: WalletCapabilities;
  /** The chainId for the transaction */
  chainId: number;
  /** The child components to be rendered within the provider component */
  children: ReactNode;
  /**
   * @deprecated Use `calls` instead.
   */
  contracts?: Calls | Contracts | Array<Call | ContractFunctionParameters>;
  /** Whether the transactions are sponsored (default: false) */
  isSponsored?: boolean;
  /** An optional callback function that handles errors within the provider */
  onError?: (e: TransactionError) => void;
  /** An optional callback function that exposes the component lifecycle state */
  onStatus?: (lifecycleStatus: LifecycleStatus) => void;
  /** An optional callback function that exposes the transaction receipts */
  onSuccess?: (response: TransactionResponse) => void;
  /** An optional time (in ms) after which to reset the component */
  resetAfter?: number;
};

/**
 * Note: exported as public Type
 */
export type TransactionReact = {
  /** An array of calls to be made in the transaction */
  calls?: Calls | Contracts | Array<Call | ContractFunctionParameters>;
  /**
   * @deprecated Use `isSponsored` instead.
   */
  capabilities?: WalletCapabilities;
  /** The chainId for the transaction */
  chainId?: number;
  /** The child components to be rendered within the transaction component */
  children?: ReactNode;
  /** An optional CSS class name for styling the component */
  className?: string;
  /**
   * @deprecated Use `calls` instead.
   */
  contracts?: Calls | Contracts | Array<Call | ContractFunctionParameters>;
  /** Whether the transactions are sponsored (default: false) */
  isSponsored?: boolean;
  /** An optional callback function that handles transaction errors */
  onError?: (e: TransactionError) => void;
  /** An optional callback function that exposes the component lifecycle state */
  onStatus?: (lifecycleStatus: LifecycleStatus) => void;
  /** An optional callback function that exposes the transaction receipts */
  onSuccess?: (response: TransactionResponse) => void;
  /** An optional time (in ms) after which to reset the component */
  resetAfter?: number;
} & (
  | {
      children: ReactNode;
      /** An optional prop to disable submit button. Only available when children are not provided. */
      disabled?: never;
    }
  | {
      children?: never;
      /** An optional prop to disable submit button. Only available when children are not provided. */
      disabled?: boolean;
    }
);

/**
 * Note: exported as public Type
 */
export type TransactionResponse = {
  transactionReceipts: TransactionReceipt[];
};
// ...
export type WalletCapabilities = {
  paymasterService?: PaymasterService;
};